Converted test/test.js to integration.test.js
